move phenotype code
